+++ lib/ts/recipe/emailpassword/types.ts
@@ -250,6 +250,12 @@ export interface RecipeInterface {
      * @deprecated Please do not override this function
      *   */
     getUserCount(): Promise<number>;
+
+    updateEmailOrPassword(input: {
+        userId: string;
+        email?: string;
+        password?: string;
+    }): Promise<{ status: "OK" | "UNKNOWN_USER_ID" | "EMAIL_ALREADY_EXISTS_ERROR" }>;
 }
 
 export type APIOptions = {