+++ lib/ts/recipe/session/index.ts
@@ -69,7 +69,7 @@ export default class SessionWrapper {
         return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.revokeMultipleSessions({ sessionHandles });
     }
 
-    /** @deprecated Use getSessionInformation instead **/
+    /** @deprecated Use getSessionInformation instead IF using core version >= 3.5 **/
     static getSessionData(sessionHandle: string) {
         return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getSessionData({ sessionHandle });
     }
@@ -81,7 +81,7 @@ export default class SessionWrapper {
         });
     }
 
-    /** @deprecated Use getSessionInformation instead **/
+    /** @deprecated Use getSessionInformation instead IF using core version >= 3.5 **/
     static getJWTPayload(sessionHandle: string) {
         return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getJWTPayload({ sessionHandle });
     }
@@ -109,12 +109,12 @@ export let revokeSession = SessionWrapper.revokeSession;
 
 export let revokeMultipleSessions = SessionWrapper.revokeMultipleSessions;
 
-/** @deprecated Use getSessionInformation instead **/
+/** @deprecated Use getSessionInformation instead IF using core version >= 3.5 **/
 export let getSessionData = SessionWrapper.getSessionData;
 
 export let updateSessionData = SessionWrapper.updateSessionData;
 
-/** @deprecated Use getSessionInformation instead **/
+/** @deprecated Use getSessionInformation instead IF using core version >= 3.5 **/
 export let getJWTPayload = SessionWrapper.getJWTPayload;
 
 export let updateJWTPayload = SessionWrapper.updateJWTPayload;