+++ lib/ts/recipe/thirdpartyemailpassword/index.ts
@@ -63,6 +63,13 @@ export default class Wrapper {
         return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUserById({ userId });
     }
 
+    static getUsersByEmail(email: string) {
+        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUsersByEmail({ email });
+    }
+
+    /**
+     * @deprecated Use supertokens.getUsersByEmail(...) function instead IF using core version >= 3.5
+     *   */
     static getUserByEmail(email: string) {
         return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUserByEmail({ email });
     }
@@ -96,6 +103,10 @@ export default class Wrapper {
         return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUserCount();
     }
 
+    static updateEmailOrPassword(input: { userId: string; email?: string; password?: string }) {
+        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.updateEmailOrPassword(input);
+    }
+
     static createEmailVerificationToken(userId: string) {
         return Recipe.getInstanceOrThrowError().createEmailVerificationToken(userId);
     }
@@ -106,6 +117,14 @@ export default class Wrapper {
 
     static isEmailVerified(userId: string) {
         return Recipe.getInstanceOrThrowError().isEmailVerified(userId);
+    }
+
+    static async revokeEmailVerificationTokens(userId: string): Promise<void> {
+        await Recipe.getInstanceOrThrowError().revokeEmailVerificationTokens(userId);
+    }
+
+    static async unverifyEmail(userId: string): Promise<void> {
+        await Recipe.getInstanceOrThrowError().unverifyEmail(userId);
     }
 
     static Google = thirdPartyProviders.Google;
@@ -131,7 +150,12 @@ export let getUserById = Wrapper.getUserById;
 
 export let getUserByThirdPartyInfo = Wrapper.getUserByThirdPartyInfo;
 
+/**
+ * @deprecated Use supertokens.getUsersByEmail(...) function instead IF using core version >= 3.5
+ *   */
 export let getUserByEmail = Wrapper.getUserByEmail;
+
+export let getUsersByEmail = Wrapper.getUsersByEmail;
 
 export let createResetPasswordToken = Wrapper.createResetPasswordToken;
 
@@ -142,6 +166,10 @@ export let createEmailVerificationToken = Wrapper.createEmailVerificationToken;
 export let verifyEmailUsingToken = Wrapper.verifyEmailUsingToken;
 
 export let isEmailVerified = Wrapper.isEmailVerified;
+
+export let revokeEmailVerificationTokens = Wrapper.revokeEmailVerificationTokens;
+
+export let unverifyEmail = Wrapper.unverifyEmail;
 
 /**
  * @deprecated Use supertokens.getUsersOldestFirst(...) function instead IF using core version >= 3.5
@@ -157,6 +185,8 @@ export let getUsersNewestFirst = Wrapper.getUsersNewestFirst;
  * @deprecated Use supertokens.getUserCount(...) function instead IF using core version >= 3.5
  *   */
 export let getUserCount = Wrapper.getUserCount;
+
+export let updateEmailOrPassword = Wrapper.updateEmailOrPassword;
 
 export let Google = Wrapper.Google;