+++ lib/ts/recipe/thirdpartyemailpassword/recipeImplementation/index.ts
@@ -61,6 +61,20 @@ export default class RecipeImplementation implements RecipeInterface {
         return await this.thirdPartyImplementation.getUserById(input);
     };
 
+    getUsersByEmail = async ({ email }: { email: string }): Promise<User[]> => {
+        let userFromEmailPass: User | undefined = await this.emailPasswordImplementation.getUserByEmail({ email });
+
+        if (this.thirdPartyImplementation === undefined) {
+            return userFromEmailPass === undefined ? [] : [userFromEmailPass];
+        }
+        let usersFromThirdParty: User[] = await this.thirdPartyImplementation.getUsersByEmail({ email });
+
+        if (userFromEmailPass !== undefined) {
+            return [...usersFromThirdParty, userFromEmailPass];
+        }
+        return usersFromThirdParty;
+    };
+
     getUserByThirdPartyInfo = async (input: {
         thirdPartyId: string;
         thirdPartyUserId: string;
@@ -79,6 +93,9 @@ export default class RecipeImplementation implements RecipeInterface {
         return userInfo.email;
     };
 
+    /**
+     * @deprecated Please do not override this function
+     *   */
     getUserByEmail = async (input: { email: string }): Promise<User | undefined> => {
         return this.emailPasswordImplementation.getUserByEmail(input);
     };
@@ -178,5 +195,13 @@ export default class RecipeImplementation implements RecipeInterface {
         let promise1 = this.emailPasswordImplementation.getUserCount();
         let promise2 = this.thirdPartyImplementation !== undefined ? this.thirdPartyImplementation.getUserCount() : 0;
         return (await promise1) + (await promise2);
+    };
+
+    updateEmailOrPassword = async (input: {
+        userId: string;
+        email?: string;
+        password?: string;
+    }): Promise<{ status: "OK" | "UNKNOWN_USER_ID" | "EMAIL_ALREADY_EXISTS_ERROR" }> => {
+        return this.emailPasswordImplementation.updateEmailOrPassword(input);
     };
 }