+++ lib/ts/recipe/thirdpartyemailpassword/types.ts
@@ -196,6 +196,8 @@ export type TypeNormalisedInput = {
 export interface RecipeInterface {
     getUserById(input: { userId: string }): Promise<User | undefined>;
 
+    getUsersByEmail(input: { email: string }): Promise<User[]>;
+
     getUserByThirdPartyInfo(input: { thirdPartyId: string; thirdPartyUserId: string }): Promise<User | undefined>;
 
     /**
@@ -250,6 +252,9 @@ export interface RecipeInterface {
         password: string;
     }): Promise<{ status: "OK"; user: User } | { status: "WRONG_CREDENTIALS_ERROR" }>;
 
+    /**
+     * @deprecated Please do not override this function
+     *   */
     getUserByEmail(input: { email: string }): Promise<User | undefined>;
 
     createResetPasswordToken(input: {
@@ -260,6 +265,12 @@ export interface RecipeInterface {
         token: string;
         newPassword: string;
     }): Promise<{ status: "OK" | "RESET_PASSWORD_INVALID_TOKEN_ERROR" }>;
+
+    updateEmailOrPassword(input: {
+        userId: string;
+        email?: string;
+        password?: string;
+    }): Promise<{ status: "OK" | "UNKNOWN_USER_ID" | "EMAIL_ALREADY_EXISTS_ERROR" }>;
 }
 
 export type EmailPasswordAPIOptions = EmailPasswordAPIOptionsOriginal;