+++ CHANGELOG.md
@@ -62,6 +62,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   `Session.getSessionData()`, `Session.getJWTPayload()`
 -   Adds email verification function calls in thirdparty sign in up API as per https://github.com/supertokens/supertokens-core/issues/295
 -   Adds `emailVerificationRecipeImplementation` in all auth recipe `APIOptions` so that APIs can access the email verification implementation.
+-   Add recipe function to fetch third party users https://github.com/supertokens/supertokens-core/issues/277
+-   Deprecates `getUserByEmail` in thirdpartyemailpassword and replaces it with `getUsersByEmail`.
+-   Adds `updateEmailOrPassword` recipe function to emailpassword and thirdpartyemailpassword recipes: https://github.com/supertokens/supertokens-core/issues/275
+-   Adds emailverification recipe functions to all recipes: https://github.com/supertokens/supertokens-core/issues/270
 
 ## [6.0.4] - 2021-07-29